DozeSensors: only use proximity sensor if supported
On msm-4.14 devices, when the proximity sensor is in use,
the smp2p-sleepstate IRQ is fired multiple times a second,
with each one holding a 200ms wakelock.
This is probably a bug in the DSP firmware.
To fix this, avoid using the proximity sensor in doze mode,
because sleep is preferred to turning off the screen.

Change-Id: I57750afd77267abdc49780f70636626d20e666ad

fix up to: DozeSensors: only use proximity sensor if supported

With the addition of the secondary sensor, not only
will we not register the sensor to prevent a null,
using the same variable also bypass the call to
the proximity sensor

Change-Id: I217f45955d50b20e617c0c7f37f9b703811edc1d
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>