SystemUI: screenshot: open long screenshot activity for partial screenshots
The partial screenshot UI is basically unusable. Replace it
with the long  screenshot UI which offers cropping, saving,
canceling, and sharing capabilities.

Change-Id: I27f7a811838f7e3f2057f567c7a350255996a8cb

-----

SystemUI: screenshot: remove duplicate clamping on move action

On MotionEvent.ACTION_DOWN, a range of of motion is calculated and
stored. On MotionEvent.ACTION_MOVE, the new position of the crop
boundary is clamped by the original range of motion, and also
clamped again by a new range of motion inside setBoundaryPosition().

The original range of motion is pointless and only used in this case,
remove it.

Change-Id: Ib4f35bf5317f5b5879606680beee3e077f829d10

-----

SystemUI: screenshot: add extra crop boundaries

Add corner boundaries that allow resizing in both vertical and
horizontal directions at the same time.

To handle corner boundaries, split every boundary into vertical
and horizontal boundaries, and handle both vertical and horizontal
changes for all boundaries. If the boundary is a purely vertical
(top, bottom) or purely horizontal (left, right) boundary, the
opposite direction boundary will be none, and the selection won't
be changed on that selection.

Also, add a middle boundary that allows moving the current selection.

To achieve this, treat the middle boundary as composed of top and
left boundaries, and when trying to change either the top or the
left boundary while the currently dragged boundary is the middle
one, reposition the opposite boundary too.

Change-Id: I6dd703733b3d161f5979d91a124350fb1a8d52ae

-----

SystemUI: screenshot: disable magnification for partial screenshots

Partial screenshots are screen-size and don't need a magnifying glass.

Change-Id: I472103352411d1482423230a4ff08ddda45cc53d

-----

SystemUI: screenshot: close QS after launching long screenshot activity

Change-Id: Ief5510c037b27be714f8f6cc8661d81931f3babd
Co-authored-by: LuK1337 <priv.luk@gmail.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>