fw/b: Introduce dynamically themed bootanimation
[Waxaranai]: Adapt code to A13

A majority of the code was already merged into AOSP during the R16/QPR1 merge,
what was missing was interaction with ThemeOverlayController to actually set
these props for them to be utilised.

Here's a brief explaination of how it works:
Monet theme is generated. and mConfigurationListener in ThemeOverlayController
detects the overlay change, this leads to it calling setBootColorProps which sets
4 props (persist.bootanim.color{1, 2, 3, 4)} These props are what store the
colors, they get colors from all 3 accent categories (Primary, Secondary and
Tertiary). setBootColorProps is also called at the end of start (void start())
to avoid actually needing to regenerate overlays for people that like the default
wallpaper.

Note: Making a bootanimation which works with this IS NOT easy. You'd need
to mess with opacity and colors of all sorts to get these colored and then
also take care of the aliasing and sharp edges with extra pixels, not for
the faint hearted.

Change-Id: Ic095ec6a55b97ac2943e5eb079068b271633547a
Signed-off-by: Maitreya29 <maitreyapatni30@gmail.com>
Signed-off-by: Waxaranai <waxaranai@gmail.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>